requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
mysql-connector-python
python-dotenv
